+++ tests/test_user_auth.py
@@ -2,7 +2,8 @@
 import json
 import sys
 
-from mock import Mock, patch
+import requests
+from mock import patch
 from nose.tools import raises, assert_equals, assert_dict_equal
 from requests.exceptions import HTTPError
 
@@ -13,40 +14,25 @@ class TestUserAuth(object):
     """Test class fo user auth. This uses Mock to mock external API calls"""
     @classmethod
     def setup_class(cls):
-        cls.client = UserAuth(
-                        "https://cerberus.fake.com",
-                        'testuser',
-                        'hardtoguesspasswd')
-        cls.auth_resp = """{"status": "mfa_req", "data":
-                        {"username": "unicorn@rainbow.com",
-                        "state_token": "0127a384d305138d4e",
-                        "client_token": "None", "user_id": "1325",
-                        "devices": [{"id": "223", "name":
-                        "Google Authenticator"}]}}"""
+        cls.client = UserAuth("https://cerberus.fake.com", 'testuser', 'hardtoguesspasswd')
+        cls.auth_resp = {
+            "status": "mfa_req",
+            "data": {
+                "username": "unicorn@rainbow.com",
+                "state_token": "0127a384d305138d4e",
+                "client_token": "None", "user_id": "1325",
+                "devices": [{"id": "223", "name": "Google Authenticator"}]
+            }
+        }
 
-    """
-    modeled after https://goo.gl/WV2WGe
-    """
-    def _mock_response(
-            self,
-            status=200,
-            content="STUFF",
-            json_data=None,
-            text="""{"key": "value"}""",
-            raise_for_status=None):
-
-        mock_resp = Mock()
-        mock_resp.raise_for_status = Mock()
-        if raise_for_status:
-            mock_resp.raise_for_status.side_effect = raise_for_status
+    @staticmethod
+    def _mock_response(status=200, reason=None, content=None):
+        mock_resp = requests.Response()
         mock_resp.status_code = status
-        mock_resp.content = content
-        mock_resp.text = text
-        # add json data if provided
-        if json_data:
-            mock_resp.json = Mock(
-                return_value=json_data
-            )
+        # Reason the status code occurred.
+        mock_resp.reason = reason
+        # Raw content in byte
+        mock_resp._content = bytes(content)
         return mock_resp
 
     def test_username(self):
@@ -68,12 +54,12 @@ class TestUserAuth(object):
     @patch('requests.get')
     def test_get_auth(self, mock_get):
         # mock return response
-        mock_resp = self._mock_response(text=self.auth_resp)
+        mock_resp = self._mock_response(content=json.dumps(self.auth_resp))
         mock_get.return_value = mock_resp
         response = self.client.get_auth()
 
         # confirm response matches the mock
-        assert_dict_equal(response, json.loads(self.auth_resp))
+        assert_dict_equal(response, self.auth_resp)
 
     if sys.version_info[0] < 3:
         input_module = '__builtin__.input'
@@ -83,39 +69,38 @@ class TestUserAuth(object):
     @patch(input_module, return_value='0987654321')
     @patch('requests.post')
     def test_mfa_response(self, mock_post, mock_input=None):
-        mfa_data = """{
-                      "status" : "success",
-                      "data" : {
-                        "user_id" : "134",
-                        "username" : "unicorn@rainbow.com",
-                        "state_token" : null,
-                        "devices" : [ ],
-                        "client_token" : {
-                          "client_token" : "61e3-f3f-6536-a3e6-b498161d",
-                          "policies" : [ "cloud-events-owner", "pixie-dust-owner"],
-                          "metadata" : {
-                            "groups" : "Rainbow.Playgroun.User,CareBear.users",
-                            "is_admin" : "false",
-                            "username" : "unicorn@rainbow.com"
-                          },
-                          "lease_duration" : 3600,
-                          "renewable" : true
-                        }
-                      }
-                    }"""
-        # mock all the things
-        mock_post.return_value = Mock()
-        mock_post.return_value.status_code = 200
-        mock_post.return_value.text = mfa_data
+        mfa_data = {
+            "status": "success",
+            "data": {
+                "user_id": "134",
+                "username": "unicorn@rainbow.com",
+                "state_token": None,
+                "devices": [],
+                "client_token": {
+                    "client_token": "61e3-f3f-6536-a3e6-b498161d",
+                    "policies": ["cloud-events-owner", "pixie-dust-owner"],
+                    "metadata": {
+                        "groups": "Rainbow.Playgroun.User,CareBear.users",
+                        "is_admin": "false",
+                        "username": "unicorn@rainbow.com"
+                    },
+                    "lease_duration": 3600,
+                    "renewable": True
+                }
+            }
+        }
 
-        response = self.client.get_mfa(json.loads(self.auth_resp))
+        # mock all the things
+        mock_post.return_value = self._mock_response(content=json.dumps(mfa_data))
+
+        response = self.client.get_mfa(self.auth_resp)
 
         # confirm the json matches
-        assert_dict_equal(response, json.loads(mfa_data))
+        assert_dict_equal(response, mfa_data)
 
     @raises(HTTPError)
     @patch('requests.get')
     def test_when_not_200_status_code(self, mock_get):
-        mock_resp = self._mock_response(status=404, raise_for_status=HTTPError("google is down"))
+        mock_resp = self._mock_response(status=404, reason='Not Found')
         mock_get.return_value = mock_resp
         self.client.get_auth()