+++ CHANGELOG.md
@@ -4,4 +4,8 @@ All notable changes to `Cerberus Python Client` will be documented in this file.
 
 ### Added
 
+###### 27-02-2017
+- Added the ability for setting role ARN and region that differs from default for IAM role auth.
+
+###### 23-02-2017
 - Initial open source code drop for the Cerberus Python client by Ann Wallace.